fix: extended timeout for the request responsible for running the script
The lytehosting.com server was killing the running process when
the connection was dropped.

fix: corrected error when WHMCS sent an email with the installation
data in case of generating an invalid character in the password

When a special character such as "{" was in the generated password,
the template (smarty) for the email could not compile and threw an error

fix: fixed stopping the script in case of
restarting when installation was already in progress.